### 2019-09-16 #1733 @whedon
[PRE REVIEW]: GraphColoring: A Java package for solving the graph coloring problem
**Submitting author:** @shah314 (<a href="http://orcid.org/0000-0002-3770-1391">Shalin S Shah</a>)
**Repository:** <a href="https://github.com/shah314/graphcoloring" target ="_blank">https://github.com/shah314/graphcoloring</a>
**Version:** v1.1
**Reviewers:** HaoZeke
