### 2019-09-16 4ce7b0f
Update paper.bib
