+++ paper.bib
@@ -9,13 +9,8 @@
   publisher={ACM}
 }
 
-@article{brelaz1979new,
-  title={New methods to color the vertices of a graph},
-  author={Br{\'e}laz, Daniel},
-  journal={Communications of the ACM},
-  volume={22},
-  number={4},
-  pages={251--256},
-  year={1979},
-  publisher={ACM}
+@article{culberson1992iterated,
+  title={Iterated greedy graph coloring and the difficulty landscape},
+  author={Culberson, Joseph},
+  year={1992}
 }
