### 2019-09-17 #1733 @jedbrown
@shah314 Could you please add a comparison with existing graph coloring capability (why would a practitioner use this software instead of others, and when should they choose others?) and address [community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) and automated testing?
