### 2019-09-17 cb1a8f0
Update paper.bib
