+++ paper.bib
@@ -14,3 +14,14 @@
   author={Culberson, Joseph},
   year={1992}
 }
+
+@article{malaguti2010survey,
+  title={A survey on vertex coloring problems},
+  author={Malaguti, Enrico and Toth, Paolo},
+  journal={International transactions in operational research},
+  volume={17},
+  number={1},
+  pages={1--34},
+  year={2010},
+  publisher={Wiley Online Library}
+}
