+++ paper.md
@@ -25,6 +25,8 @@ bibliography: paper.bib
 
 The graph coloring problem aims at assigning colors to the nodes of a graph such that no two connected nodes have the same color. The graph coloring problem is NP-complete and one of the harder problems to solve. We present a heuristic to solve it using three cascaded algorithms. The graph coloring problem was one of Karp's 21 NP-complete problems, and is also know as the problem of finding the chromatic number of a graph. Several other problems reduce to graph coloring including solving generalized Sudoku puzzles.
 
+The problem of graph coloring can be solved using exact approaches like branch and cut and using heuristics like this work. Exact algorithms are not suitable for very large instances like more than 500 vertices. This paper [malaguti2010survey] surveys some of the algorithms that are exact and heuristic and find that the %gap for exact algorithms, even when allowed to run for 7200 seconds is quite large. This algorithm is able to solve several large instances, and finds the optimum chromatic number or at least a very close to optimum solution in a short amount of computational duration. When an exact algorithm is available and suitable, it is of course the preferred way as the solution can be provably optimal. But for very large graphs, this is often unsuitable because of the exponential complexity of NP-hard problems. Our heuristic can help in these cases. Results of our algorithm can be found on our git page (<https://github.com/shah314/graphcoloring>).
+
 This Java code uses the DSATUR [@brelaz1979new] heuristics along with iterated greedy heuristics [@culberson1992iterated] to color a graph. The DSATUR heuristic orders the nodes of a graph in non-increasing order of the degree of saturation. The degree of saturation is the number of colors found in the adjacency list of a node. The iterated greedy heuristics perform randomized ordering of the vertices to color them in that order. By randomizing this process, improvement can be found in successive iterations. The algorithm then uses min-conflicts local search to improve the coloring. The method is quite successful in finding good colorings of the majority of the publicly available data sets. Results can be found on the git page.
 
 The method uses the following steps:
