### 2019-09-17 241fe0c
Update paper.md
