### 2019-09-17 #1733 @shah314
@jedbrown Thank you. I am working on it. About community guidelines, where do I add them? I think the repository is open for anyone to contribute and add issues, pull requests etc. I will add tests in a day or two.
