### 2019-09-17 #1733 @jedbrown
Usually people say something about support channels and expectations in their README and add [CONTRIBUTING.md](https://help.github.com/en/articles/setting-guidelines-for-repository-contributors) and possibly a code of conduct. See also the [Community tab](https://github.com/shah314/graphcoloring/community) for your project.
