### 2019-09-17 005cadd
Adding test bash script
Adding test bash script
