### 2019-09-17 55641b8
Create CONTRIBUTING.md
