### 2019-09-17 2bbb2bc
Create CODE_OF_CONDUCT.md
