### 2019-09-17 #1733 @shah314
@jedbrown I have added community guidelines, a code of conduct and a test case to test the code. Since this is java code, it has to be compiled, so a Java compiler is a requirement. I have added a tests folder and a test.sh script which will compile the code and run a test. Please see if you are able to run this script. I also added a README to the tests folder.
