### 2019-09-17 45a65d9
Update test.sh
