+++ tests/test.sh
@@ -5,6 +5,8 @@ cp ../../*.java .
 javac *.java
 cp ../data.col .
 java GraphColoring
-rm *
 cd ..
+rm testing1/*.java
+rm testing1/*.class
+rm testing1/data.col
 rmdir testing1
