### 2019-09-17 #1733 @shah314
@jedbrown I have also made some changes to the paper with some motivation on why a heuristic might be preferred to an exact algorithm (like branch and cut).
