### 2019-09-17 #1733 @jedbrown
Thanks. In terms of comparison, I was thinking in terms of existing software such as [ColPack](https://cscapes.cs.purdue.edu/coloringpage/software.htm), the classic MINPACK methods ([largest first](https://www.mcs.anl.gov/petsc/petsc-current/docs/manualpages/MatOrderings/MATCOLORINGLF.html#MATCOLORINGLF), [smallest last](https://www.mcs.anl.gov/petsc/petsc-current/docs/manualpages/MatOrderings/MATCOLORINGSL.html#MATCOLORINGSL), and [incidence degree](https://www.mcs.anl.gov/petsc/petsc-current/docs/manualpages/MatOrderings/MATCOLORINGID.html#MATCOLORINGID) -- available in PETSc, for example), and implementations of parallel methods such as [Jones-Plassmann](https://www.mcs.anl.gov/petsc/petsc-current/docs/manualpages/MatOrderings/MATCOLORINGJP.html) and [Bozdag's greedy method](https://www.mcs.anl.gov/petsc/petsc-current/docs/manualpages/MatOrderings/MATCOLORINGGREEDY.html). Does your implementation outperform these? Is it significant because it's more convenient for a Java caller? These are the sort of questions that you can answer in your documentation and paper.
