### 2019-09-17 #1733 @shah314
@jedbrown Thanks. As far as I can see, ColPack does not publish benchmarks on publicly available data sets. The other references you cited all use their own data which are not publicly available. The paper I have cited uses publicly available benchmarks (like fpsol, myciel and queen) which can be verified easily. So, I'm not sure how to compare. I will keep looking. Thanks.
