### 2019-09-17 #1733 @jedbrown
These are open source packages that implement graph coloring as part of solving higher level problems. It should be straightforward to run on existing data sets. (Thousands of people use these implementations every day, but I have not seen their results on the benchmarks you mention.)

Note that your CMU link has an invalid certificate. Perhaps you can ask the owners to fix that.
