### 2019-09-18 #1733 @shah314
@jedbrown I'm trying to run the algorithms you mentioned on PETSc. I was able to set up PETSc and configure and build it. But I am unable to figure out how to invoke it and run these algorithms. When I go to the src directory and try to build them using make, it does not work because there is only a lib target. Any help?
