### 2019-09-19 #1733 @shah314
@jedbrown I was able to use ColPack to see how the largest first, smallest last and the incidence degree algorithms do on my instances. I have the results at the bottom of the git README. I have also included a summary of the comparison in the paper. Thanks.
