### 2019-09-19 9ec9502
Convert DIMACS col files to ColPack mtx files
Convert DIMACS col files to ColPack mtx files
