### 2019-09-19 #1733 @shah314
Also, if you are unable to download individual instance files from the CMU site, try downloading the tar file. I have also uploaded the instances tar file on my repository: https://github.com/shah314/graphcoloring/blob/master/instances.tar
