### 2019-09-19 31983e1
Update test.sh
