+++ paper.bib
@@ -25,3 +25,10 @@
   year={2010},
   publisher={Wiley Online Library}
 }
+
+@article{gebremedhin2010colpack,
+  title={ColPack: Graph coloring software for derivative computation and beyond},
+  author={Gebremedhin, AH and Nguyen, Duc and Patwary, MMA and Pothen, A},
+  journal={Submitted to ACM TOMS},
+  year={2010}
+}
