+++ README.md
@@ -23,17 +23,19 @@ Implementation of the DSatur[1] heuristics for <A href="http://en.wikipedia.org/
 Instances are available <a href="http://mat.gsia.cmu.edu/COLOR/instances.html">here</a>
 and <a href="http://www.nlsde.buaa.edu.cn/~kexu/benchmarks/graph-benchmarks.htm">here</a>.</p>
 
-Compile the code and then run:
-<pre>
-java GraphColoring
-</pre>
 
-<ul>
-<li>Change FILE in Constants.java to the file that contains the graph in a known format.
-(defaults to 'data.col' in the current directory)</li>
-<li>Please remove all comments (lines starting with a 'c') and other extraneous text from the file.</li>
-<li>Compile the code and run "java GraphColoring".</li>
-</ul>
+<pre>
+Compile the code and then run:
+
+java GraphColoring (for the heuristics)
+OR
+java Backtracking (for backtracking)
+
+Change FILE in Constants.java to the file that contains the graph in a known format.
+(defaults to 'data.col' in the current directory)
+
+Please remove all comments (lines starting with a 'c') and other extraneous text from the file.
+</pre>
 
 <p>
 [1] "New methods to color the vertices of a graph", Brelaz D., CACM 22(4) pp 251--256.<br>
