### 2019-09-20 c567f24
Update paper.md
