### 2019-09-20 #1733 @shah314
I have fixed the invalid certificate error for the CMU link.
