### 2019-09-25 e562550
Update test.sh
