### 2019-09-25 183543b
Update testbacktracking.sh
