+++ tests/testbacktracking.sh
@@ -5,8 +5,3 @@ cp ../../*.java .
 javac *.java
 cp ../data.col .
 java Backtracking
-cd ..
-rm testing1/*.java
-rm testing1/*.class
-rm testing1/data.col
-rmdir testing1
