+++ GraphColoring.java
@@ -8,20 +8,28 @@ import java.util.*;
  * Email: shah.shalin@gmail.com
  */
 public class GraphColoring {
-    
+
     /** Creates a new instance of GraphColoring */
     public GraphColoring() {
     }
-    
+
     public static void main(String [] args) throws Exception
     {
+        if(args.length == 0)
+        {
+            System.out.println("Usage: java GraphColoring <filename>");
+            System.out.println("Example: java GraphColoring data.col");
+            System.exit(0);
+        }
+
+        Constants.FILE = args[0];
         System.out.println("Reading Graph...");
         /* Read the graph from Constants.FILE */
         Graph graph = GraphReader.readGraph();
-        
+
         /* Compute Clique */
         LinkedHashSet clique = MaxClique.computeClique(graph);
-        
+
         /* Color the vertices of the clique with different colors */
         int k = clique.size();
         Iterator it = clique.iterator();
@@ -33,7 +41,7 @@ public class GraphColoring {
             node.colorNode(col);
             col++;
         }
-            
+
         Node [] nodes = graph.nodes;
         PossibleColorsComparator comparator = new PossibleColorsComparator();
         TreeSet uncoloredNodes = new TreeSet(comparator);
@@ -50,7 +58,7 @@ public class GraphColoring {
                 flags.add(new Integer(node.value));
             }
         }
-        
+
         while(!uncoloredNodes.isEmpty())
         {
             Node node = (Node)uncoloredNodes.first();
@@ -68,15 +76,15 @@ public class GraphColoring {
                     {
                         int vertex = ((Integer)it.next()).intValue();
                         Node n = graph.nodes[vertex];
-                        
+
                         if(uncoloredNodes.contains(n))
                         {
-                            uncoloredNodes.remove(n);       
+                            uncoloredNodes.remove(n);
                             n.computeDegreeSat(graph);
                             uncoloredNodes.add(n);
                         }
                     }
-                    
+
                     if(i > k)
                     {
                         k = i;
@@ -85,10 +93,10 @@ public class GraphColoring {
                 }
             }
         }
-        
+
         System.out.println(k + " coloring found using DSatur.");
         System.out.println("Applying Iterated Greedy Improvement...");
-        
+
         int [] colors = IteratedGreedy.iteratedGreedy(k, graph);
         int maxColor = -1;
         for(int i=0; i<colors.length; i++)
@@ -103,7 +111,7 @@ public class GraphColoring {
                 maxColor = colors[i];
             }
         }
-        
+
         System.out.println("Applying Local Search...");
         colors = LocalSearch.localSearch(graph, maxColor);
         maxColor = -1;
@@ -118,7 +126,7 @@ public class GraphColoring {
                 maxColor = colors[i];
             }
         }
-        
+
         System.out.println("Final Coloring of graph possible with " + maxColor + " colors.");
         System.out.println("Colors of Vertices: ");
         for(int i=0; i<colors.length; i++)
@@ -126,19 +134,19 @@ public class GraphColoring {
             System.out.print(colors[i] + " ");
         }
     }
-    
+
     public static class PossibleColorsComparator implements Comparator
     {
         public int compare(Object o1, Object o2)
         {
             Node node1 = (Node)o1;
             Node node2 = (Node)o2;
-            
+
             if(node1.value == node2.value)
             {
                 return 0;
             }
-            
+
             if(node1.degreeSat < node2.degreeSat)
             {
                 return 1;
