+++ tests/README.md
@@ -5,7 +5,7 @@ Please follow the following steps to test the code:
 <ol>
 <li>Use the data.col (queen9_9) file in this directory and copy it to the directory where the source code is</li>
 <li>Compile the Java code</li>
-<li>Run "java GraphColoring" for the heuristic algorithms (test.sh)</li>
+<li>Run "java GraphColoring data.col" for the heuristic algorithms (test.sh)</li>
 <li>The algorithm should then follow a cascaded run of various steps:
 <pre>
 Reading Graph...
@@ -25,5 +25,5 @@ Colors of Vertices:
 2 4 1 5 9 11 7 8 6 9 10 6 7 2 ...</pre>
 </li>
 <li>The coloring should be around 11 (it might vary as the algorithm is randomized)</li>
-<li>(Alternatively, run "java Backtracking" for the backtracking algorithm (testbacktracking.sh))</li>
+<li>(Alternatively, run "java Backtracking data.col" for the backtracking algorithm (testbacktracking.sh))</li>
 </ol>
