### 2019-09-26 #1733 @shah314
Any further thoughts, @jedbrown ?

I have made some changes to the code and added tests. I am now passing in the file name to the code, so there is no need for the data.col to exist in the current directory.

Please do a git pull and review and verify the changes.
