### 2019-09-30 #1733 @jedbrown
@shah314 Thanks for your diligence in comparing with ColPack. I just remembered that NetworkX also has coloring algorithms, though I don't think they are any better than those in ColPack.

In your README, it just says "compile the code". I think it would be helpful to list the command to run to do that. From the tests, you evidently use the following, which gives some warnings with OpenJDK-12.

```
$ javac *.java
Note: Some input files use or override a deprecated API.
Note: Recompile with -Xlint:deprecation for details.
Note: Some input files use unchecked or unsafe operations.
Note: Recompile with -Xlint:unchecked for details.
```

Could you comment on the expected execution-time performance of the algorithm? For example, packages like ColPack are used to color graphs with millions of vertices in a setup step for a solver. (They are usually computing a distance-2 coloring, though distance-1 colorings are used in a related context to expose parallelism.)

You have a lot of HTML inlined into your markdown files. Is that deliberate to overcome some perceived limitation of markdown?

Would you be willing to add automated testing to your repository, or at least streamline the testing procedure so that there is one command to run at the top level that reports success/failure?

I'll look for reviewers in the meantime.
