### 2019-09-30 a3f38e0
Create testall.sh
