+++ README.md
@@ -5,7 +5,10 @@
 
 <div align="center"><img src="threecoloring.jpg"/></div>
 <p>
-Implementation of the DSatur[1] heuristics for <A href="http://en.wikipedia.org/wiki/Graph_coloring_problem">graph coloring</a> in Java. The heuristic follows the following steps:
+Implementation of the DSatur[1] heuristics for <A href="http://en.wikipedia.org/wiki/Graph_coloring_problem">graph coloring</a> in Java. The algorithm is created with tha aim of obtaining the best coloring, irrespective of run time. If you need to speed up the algorithm, consider not using local search. Also, if you need help with this, please open an issue.</p>
+
+<p>
+	The heuristic follows the following steps:
 </p>
 <ol>	
 	<li>Compute a clique (maximum is good)
