### 2019-09-30 #1733 @shah314
The compilation warning is because the code does not use generics. I think it is completely harmless.

About execution time, there are several parameters in Constants.java which can be used to control the run time. For instance, one could disable local search, or reduce the number of iterations and get a much better run time. On the inithx.i.3 instance with 621 vertices, this algorithm takes 2 seconds (without local search) and ColPack takes 45 milliseconds. But if the goal is to get a good enough coloring as fast as possible, the algorithm could be changed. I have mentioned this on the git README.

I mainly use HTML for the results tables, and to change the properties of text. It just makes it more readable when viewed on a browser.

I have added testall.sh which runs both the heuristics and the backtracking algorithms. It compiles the code, and runs both algorithms on an instance with a known chromatic number (queen9_9).

Thanks @jedbrown
