+++ README.md
@@ -596,7 +596,7 @@ in the following table.</p>
 
 <br>I ran <a href="https://github.com/CSCsw/ColPack">ColPack</a> (DISTANCE_ONE) on some of the publicly available data sets for comparison. The results are in the following table. On all of the instances, our algorithm is as good or better than the ColPack implementations. On the le450_5d and the queen9_9 instances, our method is able to achieve a better coloring of the graphs. 
 
-On the largest instance (inithx.i.3) with 621 vertices, ColPack took 45 milliseconds and this algorithm (without local search) took 2 seconds. It is possible to reduce the run time by changing the variables in Constants.java. Please open an issue if you need help with reducing the run time. Local search takes the longest time, so it should be used only if you need the best coloring possible, irrespective of run time.
+About execution time, there are several parameters in Constants.java which can be used to control the run time. For instance, one could disable local search, or reduce the number of iterations and get a much better run time. On the inithx.i.3 instance with 621 vertices, this algorithm takes 2 seconds (without local search) and ColPack takes 45 milliseconds. But if the goal is to get a good enough coloring as fast as possible, the algorithm could be changed. Please open an issue if this is the case.
 <br>
 <table>
 	<tr><td><b>DataSet</b></td><td>LARGEST_FIRST</td><td>SMALLEST_LAST</td><td>INCIDENCE_DEGREE</td><td>This Algorithm</td></tr>
