### 2019-09-30 #1733 @jedbrown
Editing compile-time constants is a poor interface for a potential application. It would be *much* better to have a configurable public interface with examples. I'm not personally familiar with Java documentation practices, but it's very common for JOSS submissions to have API documentation plus tutorials written with tools like Sphinx or Doxygen. If you want the package to be used, it would be very worthwhile to create such documentation. As a lighter weight alternative, you could just write markdown files with example calling procedures.

Regarding HTML, Markdown has its own lighter weight syntax for the formatting that I've noticed. This isn't that important, but is disruptive when reading the text files directly (e.g., after cloning) and adds a layer of uncertainty for a potential contributor to determine whether to introduce inconsistency by using markdown or to reproduce via HTML to satisfy some unknown requirement (such as that it be copy-pastable into a literal HTML file).

On testing, you added the following to `tests/README.md`, but it is not mentioned in the top-level `README.md`.
> Please run <b>testall.sh</b> to run all tests.

It doesn't say how to run them (and `testall.sh` is not executable). This kind of documentation should be really explicit. It also helps for tests to have explicit PASS/FAIL reports. There shouldn't be any interpretation left to the user. Perhaps you could write some tests using JUnit, for example, which would formalize what we mean by automated tests. The following is better than nothing, but is not deterministic and it's left to the reader to interpret whether the result is a pass.
```
$ bash testall.sh
mkdir: cannot create directory ‘testing1’: File exists
Note: Some input files use or override a deprecated API.
Note: Recompile with -Xlint:deprecation for details.
Note: Some input files use unchecked or unsafe operations.
Note: Recompile with -Xlint:unchecked for details.
Reading Graph...
Computing Clique...
Maximum Clique Size Found: 9
Vertices in the Clique:
40 13 49 67 4 31 76 58 22

104 milliseconds (excluding I/O).
15 coloring found using DSatur.
Applying Iterated Greedy Improvement...
Found Better Coloring - 13
Found Better Coloring - 12
Found Better Coloring - 11
Applying Local Search...
Final Coloring of graph possible with 11 colors.
Colors of Vertices:
8 2 6 9 11 1 4 5 10 5 10 11 7 2 3 8 6 1 9 4 3 10 1 6 2 11 7 6 1 7 5 3 4 10 8 2 4 8 10 11 7 9 5 3 6 11 9 1 3 8 2 6 10 4 2 7 8 4 6 10 11 1 3 10 11 2 1 9 8 3 4 5 1 3 4 8 5 11 7 2 9 Reading Graph...
Computing Clique...
Maximum Clique Size Found: 9
Vertices in the Clique:
22 13 49 67 4 31 76 58 40

108 milliseconds (excluding I/O).
Trying to color graph with 9 colors...
Trying to color graph with 10 colors...
Trying to color graph with 11 colors...
Trying to color graph with 12 colors...
Trying to color graph with 13 colors...
Trying to color graph with 14 colors...

14 coloring found! Exiting...
```
