### 2019-09-30 b0ed230
Add automated tests
Add automated tests
