### 2019-09-30 82c27b2
Add automated tests
Add automated tests
