+++ tests/README.md
@@ -1,6 +1,6 @@
 <h1>Testing the graph coloring code</h1>
 
-Please run <b>testall.sh</b> to run all tests.
+Please run <b>testsautomated.sh</b> to run all tests.
 
 Please follow the following steps to test the code:
 
