### 2019-09-30 182c3f0
Add automated tests
