### 2019-09-30 #1733 @shah314
@jedbrown I think I have addressed all of your suggestions.

The code now has several command line arguments that can be passed in. I have mentioned this in both README.md files in the tests as well as the root directories.

I have added an "automatedtests.sh" file that will output if the tests passed for both heuristics as well as the backtracking algorithms.

I created a parameters.md file that explains all the parameters that can be passed in.

About the shell scripts being not executable, I have mentioned that the user needs to run a "chmod u+x *.sh" to run the shell scripts. This is the only way I see how to do this. I don't think git propagates the permissions when a file is uploaded.
