### 2019-10-05 #1733 @jedbrown
I don't understand what environment you're on that can't commit mode changes, but it's fully supported by Git; here's an example PR:
https://github.com/shah314/graphcoloring/pull/2
Edit: I see you merged this, but the docs talking about using `chmod` should also be updated.

I also demonstrated more idiomatic markdown formatting [here](https://github.com/jedbrown/graphcoloring/blob/jed/markdown/parameters.md), to be compared with the [HTML-in-markdown](https://github.com/shah314/graphcoloring/blob/master/parameters.md). The "for heuristics" and "for backtracking" parts should be organized more clearly. Which options apply to which executable?

I think we should talk about whether a standalone executable is the primary research utility. In my research world, a potential user would vastly prefer to call a library, passing the data in-memory, versus write it to file and call a standalone executable. If you strongly believe that standalone is the only use case that needs to be supported, we can take it to review and get opinions of reviewers, but I think there is high likelihood that they will want an in-memory interface. I see from `Tests.java` that there is a library interface (though perhaps still file-based and certainly clunky by putting numbers into strings to be parsed), but there is no API documentation. It believe it is standard to write API docs using Javadoc. The resulting generated documentation could be published using GitHub pages, for example.

When called as a library, it must be possible (and preferably default) to print nothing to the screen. Imagine, for example, that someone writes a NetworkX binding to your library and using it becomes an option to their graph coloring interfaces.

In a command-line interface, one would normally expect `--help` to give more detailed usage information. I believe it is standard (also in Java) to use a command-line parsing library such as [Apache CLI](https://commons.apache.org/proper/commons-cli/) or [args4j](http://args4j.kohsuke.org/), which helps provide useful diagnostics and improve reliability.

I'm not familiar with Java community standards, but you should investigate how a command-line utility is normally packaged/distributed. If successful, your software may find it into distribution repositories such as Debian/Ubuntu, Homebrew, and Anaconda. It makes sense to research and choose an installation model now (versus asking the user to copy files around and invoke `javac` and `java` directly).

On testing, thank you for adding the affirmative check. Do you object to setting up continuous integration or is it just an unfamiliar technology?
