+++ src/main/java/Backtracking.java
@@ -1,6 +1,12 @@
 
 import java.util.*;
 
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.CommandLineParser;
+import org.apache.commons.cli.DefaultParser;
+import org.apache.commons.cli.Option;
+import org.apache.commons.cli.Options;
+
 /**
  * Backtracking solver for the graph coloring problem.
  * Spends a specific amount of time (Constants.TIME) on trying to color the graph
@@ -21,26 +27,52 @@ public class Backtracking {
 
     public static void main(String [] args) throws Exception
     {
-    	colorGraph(args);
+    	Options options = new Options();
+    	Option o1 = new Option("f", true, "The DIMACS formatted graph file name");
+    	o1.setRequired(true);
+    	options.addOption(o1);
+    	Option o2 = new Option("t", true, "Number of milliseconds to spend on each value of k");
+    	o2.setRequired(false);
+    	options.addOption(o2);
+    	
+    	CommandLineParser parser = new DefaultParser();
+    	CommandLine cmd = parser.parse( options, args);
+    	
+    	String filename = cmd.getOptionValue("f");
+    	int milliseconds = Integer.parseInt(cmd.getOptionValue("t", "1000"));
+    	System.out.println("Reading Graph...");
+        /* Read the graph from Constants.FILE */
+        Graph graph = GraphReader.readGraph(filename);
+        
+        int [] colors = colorGraph(graph, milliseconds);
+        int maxColor = -1;
+        for(int i=0; i<colors.length; i++)
+        {
+            if(maxColor == -1)
+            {
+                maxColor = colors[i];
+            }
+            else if(maxColor < colors[i])
+            {
+                maxColor = colors[i];
+            }
+        }
+        
+        System.out.println("Final Coloring of graph possible with " + maxColor + " colors.");
+        System.out.println("Colors of Vertices: ");
+        for(int i=0; i<colors.length; i++)
+        {
+            System.out.print(colors[i] + " ");
+        }       
+        
+        System.out.println("\nFinished\n");
     }
     
-    public static int colorGraph(String [] args) throws Exception
+    public static int [] colorGraph(Graph graph, int milliseconds) throws Exception
     {
-    	if(args.length == 0)
-    	{
-    		System.out.println("Usage: java Backtracking <filename>");
-    		System.out.println("Example: java Backtracking data.col");
-    		System.exit(0);
-    	}
-
-    	ParseArguments.parse(args);
-    	Constants.FILE = args[0];
-
-        System.out.println("Reading Graph...");
-        /* Read the graph from Constants.FILE */
-        Graph graph = GraphReader.readGraph();
-
-        /* Compute Clique */
+    	Constants.TIME = milliseconds;
+    	
+    	/* Compute Clique */
         LinkedHashSet clique = MaxClique.computeClique(graph);
 
         /* Color the vertices of the clique with different colors */
@@ -169,17 +201,13 @@ public class Backtracking {
         }
         
         System.out.println("Colors of Vertices: ");
-        int maxColor = Integer.MIN_VALUE;
+        int [] colors = new int [Constants.NUMBER_NODES];
         for(int i=0; i<graph.nodes.length; i++)
         {
-            System.out.print(graph.nodes[i].color + " ");
-            if(graph.nodes[i].color > maxColor)
-            {
-            	maxColor = graph.nodes[i].color;
-            }
+        	colors[i] = graph.nodes[i].color;
         }
         
-        return maxColor;
+        return colors;
     }
 
     public static class PossibleColorsComparator implements Comparator
