+++ src/main/java/Colors.java
@@ -1,0 +1,4 @@
+
+public class Colors {
+
+}
