### 2019-10-05 01997f4
Add help to CLI
Add help to CLI
