+++ src/main/java/com/gcol/Tests.java
@@ -1,3 +1,4 @@
+package com.gcol;
 
 public class Tests {
 
