+++ src/main/java/com/gcol/Constants.java
@@ -1,3 +1,4 @@
+package com.gcol;
 
 /**
  * Constants.
