+++ src/main/java/com/gcol/GraphReader.java
@@ -1,3 +1,4 @@
+package com.gcol;
 
 import java.util.*;
 import java.io.*;
