+++ src/main/java/com/gcol/DimacsToMatrix.java
@@ -1,3 +1,4 @@
+package com.gcol;
 import java.util.*;
 import java.io.*;
 
