+++ tests/test.sh
@@ -1,7 +1,4 @@
 #!/bin/bash -v
-mkdir testing1
-cd testing1
-cp ../../*.java .
-javac *.java
-cp ../data.col .
-java GraphColoring data.col
+cd ..
+mvn package
+java -cp target/graphcoloring-1.7-jar-with-dependencies.jar com.gcol.GraphColoring -f data.col -i 1000 -l true -j 1000 -m 1000
