+++ src/main/java/com/gcol/Colors.java
@@ -1,3 +1,4 @@
+package com.gcol;
 
 public class Colors {
 
