### 2019-10-05 2817c0d
Delete parameters.md
