+++ README.md
@@ -1,38 +1,32 @@
-<h1>Heuristic for Graph Coloring</h1>
+# Heuristic for Graph Coloring
 
-<i><h3>Shalin Shah</h3></i>
-<a href="https://zenodo.org/badge/latestdoi/134313250"><img src="https://zenodo.org/badge/134313250.svg" alt="DOI"></a>
+## Shalin Shah
 
-<div align="center"><img src="threecoloring.jpg"/></div>
-<p>
-Implementation of the DSatur[1] heuristics for <A href="http://en.wikipedia.org/wiki/Graph_coloring_problem">graph coloring</a> in Java. The algorithm is created with tha aim of obtaining the best coloring, irrespective of run time. If you need to speed up the algorithm, consider not using local search. Also, if you need help with this, please open an issue.</p>
+[[threecoloring.jpg]]
 
-<p>
-	The heuristic follows the following steps:
-</p>
-<ol>	
-	<li>Compute a clique (maximum is good)
-	<li>Color the clique
-	<li>Sort the rest of the vertices in non-increasing order of the 
-	degree of saturation
-	<li>Color the vertices in the order given by 3. Also, when a vertex is 
-	colored, change the degree of saturation of the neighboring vertices
-	so that the order of coloring changes
-	<li>Improve the coloring using Iterated Greedy techniques [2]
-	<li>Improve the coloring using min-conflicts local search
-	<li>Report the coloring
-</ol>
+Implementation of the DSatur[1] heuristics for graph coloring in Java. The algorithm is created with tha aim of obtaining the best coloring, irrespective of run time. If you need to speed up the algorithm, consider not using local search. Also, if you need help with this, please open an issue.
 
-<pre>
-<b>Maven</b> is needed to compile and package the code. 
-For e.g. you might run <b>brew install maven</b> on a Mac.
+The heuristic follows the following steps:
+
+	Compute a clique (maximum is good)
+	Color the clique
+	Sort the rest of the vertices in non-increasing order of the degree of saturation
+	Color the vertices in the order given by 3. Also, when a vertex is colored, change the degree of saturation of the neighboring vertices so that the order of coloring changes
+	Improve the coloring using Iterated Greedy techniques [2]
+	Improve the coloring using min-conflicts local search
+	Report the coloring
+
+Maven is needed to compile and package the code. 
+For e.g. you might run "brew install maven" on a Mac.
 Compile the code using Maven:
-<b>mvn package</b>
+
+	mvn package
+
 (Please ignore the compilation warnings, it is because the code does not use generics)
 
 Then, run the algorithm using any of the two methods:
 
-<b>Heuristics</b>
+### Heuristics
 java -cp target/graphcoloring-1.7-jar-with-dependencies.jar com.gcol.GraphColoring -f data.col
 
 java -cp target/graphcoloring-1.7-jar-with-dependencies.jar com.gcol.GraphColoring
@@ -45,7 +39,7 @@ usage: java -cp graphcoloring-1.7-jar-with-dependencies.jar com.gcol.GraphColori
  -l <arg>   Enable local search true/false
[…]
