+++ README.md
@@ -2,7 +2,7 @@
 
 ## Shalin Shah
 
-[[threecoloring.jpg]]
+![Three Coloring](threecoloring.jpg)
 
 Implementation of the DSatur[1] heuristics for graph coloring in Java. The algorithm is created with tha aim of obtaining the best coloring, irrespective of run time. If you need to speed up the algorithm, consider not using local search. Also, if you need help with this, please open an issue.
 
