+++ CONTRIBUTING.md
@@ -1,17 +1,19 @@
-<h1>Contributing to the graph coloring code</h1>
+# Contributing to the graph coloring code
+
 These are general guidelines on how to contribute to the code.
-<h3>Bugs and Issues</h3>
-<ul>
-<li>If while running the code you find that it has an issue, please report it to the issues page.</li>
-<li>If you are able to fix the issue, please open a pull request with your changes to a new branch</li>
-<li>I will carefully monitor the issues page. Please add @shah314 in the issue so I get a notification.</li>
-</ul>
-<h3>Code Enhancements</h3>
-<ul>
-<li>Before committing your changes, please see the README in the tests folder</li>
-<li>Beforing committing, please run test.sh (you may have to change the path to the bash binary).
-<li>Please commit your code enhancements to a new branch every time</li>
-<li>After committing to a new branch, please create a pull request and add @shah314 as a reviewer.</li>
-<li>Once I have reviewed your changes, I will merge the branch to master</li>
-<li>If you would like to be a reviewer, please open an issue.</li>
-</ul>
+
+## Bugs and Issues
+  
+    If while running the code you find that it has an issue, please report it to the issues page.
+    If you are able to fix the issue, please open a pull request with your changes to a new branch.
+    I will carefully monitor the issues page. Please add @shah314 in the issue so I get a notification.
+
+## Code Enhancements
+
+    Before committing your changes, please see the README in the tests folder
+    Beforing committing, please run test.sh (you may have to change the path to the bash binary).
+    Please commit your code enhancements to a new branch every time
+    After committing to a new branch, please create a pull request and add @shah314 as a reviewer.
+    Once I have reviewed your changes, I will merge the branch to master
+    If you would like to be a reviewer, please open an issue.
+
