### 2019-10-05 #1733 @shah314
@jedbrown Thank you for the thorough review.

I have made some major changes.

The build now uses Maven (which is standard for Java code).

I have added API documentation. The user can now call the code using a Java API by adding the jar in the classpath (api.md).

I have converted all md files to markdown instead of using HTML.

I have updated the tests to use Maven instead of directly using javac.

I have added a dependency on Apache Commons CLI which will be built in to the uber jar using Maven. I have also added a "-help" and -h" command line arguments to print the help screen on the arguments. Running the code without any arguments will print the help screen.

Please do a git pull and review and verify.

Thank you.
