+++ APICallBacktracking.java
@@ -1,13 +1,12 @@
-
 import com.gcol.Backtracking;
 import com.gcol.GraphReader;
 import com.gcol.Graph;
 
-public class APICallBacktracking
+public class APICallBacktracking 
 {
 	public static void main(String [] args) throws Exception
 	{
-		Graph graph = GraphReader.readGraph("./data.col");
+		Graph graph = GraphReader.readGraph("data.col");
 		// where filename is a DIMACS formatted graph file
 		int millisecondsperK = 1000;
 		int [] colors = Backtracking.colorGraph(graph, millisecondsperK);
@@ -24,7 +23,7 @@ public class APICallBacktracking
 		        maxColor = colors[i];
 		    }
 		}
-
+		        
 		System.out.println("Final Coloring of graph possible with " + maxColor + " colors.");
 		System.out.println("Colors of Vertices: ");
 		for(int i=0; i<colors.length; i++)
