### 2019-10-06 #1733 @shah314
I changed pom.xml with the maven compiler version. I think JDK version 1.8 should work fine. Please do a git pull and try again. Thanks.
