### 2019-10-06 #1733 @jedbrown
Works, though I think you need the following to suppress a warning.
```
diff --git i/pom.xml w/pom.xml
index 2a6cb04..7563c12 100644
--- i/pom.xml
+++ w/pom.xml
@@ -30,6 +30,7 @@
             </plugin>
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
                 <configuration>
                     <source>1.8</source>
                     <target>1.8</target>
```
