### 2019-10-06 #1733 @jedbrown
@gabrielspadon @sonal-raj @camillescott @hooman650 :wave: Would you be available to review this submission for JOSS?
