### 2019-10-11 4e7a0b3
Update colors of the graph object
Update colors of the graph object
