### 2019-10-29 #1843 @whedon
[REVIEW]: JCOL: A Java package for solving the graph coloring problem
**Submitting author:** @shah314 (<a href="http://orcid.org/0000-0002-3770-1391">Shalin S Shah</a>)
**Repository:** <a href="https://github.com/shah314/graphcoloring" target ="_blank">https://github.com/shah314/graphcoloring</a>
**Version:** v1.1

@HaoZeke and @ProbShin, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
