### 2019-10-29 #1843 @jedbrown
👋🏼 @HaoZeke Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [GraphColoring repository](https://github.com/shah314/graphcoloring)). I'll be watching this thread if you have any questions.
