### 2019-11-20 #1843 @jedbrown
Yes.On Nov 20, 2019 14:40, Shalin Shah <notifications@github.com> wrote:@jedbrown Can I change the name of the paper?

—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.
