### 2019-11-20 #1843 @shah314
@jedbrown I have changed the name in the paper to the following:

JCOL: A Java package for solving the graph coloring problem
