### 2019-12-07 #1843 @HaoZeke
@jedbrown sorry, been dragging my feet because of a recent move. Will finish ASAP.
