+++ paper.bib
@@ -6,13 +6,15 @@
   number={4},
   pages={251--256},
   year={1979},
-  publisher={ACM}
+  publisher={ACM},
+  doi = {10.1145/359094.359101}
 }
 
 @article{culberson1992iterated,
   title={Iterated greedy graph coloring and the difficulty landscape},
   author={Culberson, Joseph},
-  year={1992}
+  year={1992},
+  doi={10.7939/R3M32NH6Q}
 }
 
 @article{malaguti2010survey,
@@ -23,12 +25,14 @@
   number={1},
   pages={1--34},
   year={2010},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/j.1475-3995.2009.00696.x}
 }
 
 @article{gebremedhin2010colpack,
-  title={ColPack: Graph coloring software for derivative computation and beyond},
+  title={ColPack: Software for graph coloring and related problems in scientific computing},
   author={Gebremedhin, AH and Nguyen, Duc and Patwary, MMA and Pothen, A},
   journal={Submitted to ACM TOMS},
-  year={2010}
+  year={2010},
+  doi={10.1145/2513109.2513110}
 }
