### 2019-12-07 #1843 @shah314
@HaoZeke I have fixed the DOIs. Thanks for agreeing to review.
