### 2019-12-31 shah314/graphcoloring#4 @HaoZeke
[JOSS-Review] Paper Revisions
This is part of https://github.com/openjournals/joss-reviews/issues/1843. Please address the following points:
- The results should also be moved away from the git page and into the paper itself
## Comparisons

![image](https://user-images.githubusercontent.com/4336207/71634811-30202500-2c45-11ea-8654-eb7101a9dcf6.png)
- This section should be re-written and the results must be included in the paper itself.
- A graph would be better over time
- Resource usage should be clarified

The code and github resources are good, but the write up needs to be more coherent and self-contained.
