### 2020-01-10 7cc88f5
Update paper.md
