### 2020-01-10 c974b7e
Update paper.md
