### 2020-01-10 shah314/graphcoloring#4 @shah314
@HaoZeke I think I have made the changes you requested. Please review the updated paper. I have included the ColPack comparison table in the paper. I have also included the run times for ColPack and JCOL in the paper. Thank you for the review.
