### 2020-01-10 shah314/graphcoloring#4 @shah314
@jedbrown Please review the modified paper.
