### 2020-01-10 #1843 @shah314
@HaoZeke I have made the changes you requested.

@ProbShin Any thoughts or questions?
