### 2020-02-03 #1843 @shah314
@HaoZeke I have made the changes you requested (I think). Any further thoughts?
