### 2020-02-07 #1843 @HaoZeke
@shah314 Thank you for making the changes. I have completed my review and am pleased to recommend this for publication in its current state in JOSS. @jedbrown
