### 2020-02-07 #1843 @shah314
@jedbrown Can you please update the name on the JOSS website? It still shows the old name.

https://joss.theoj.org/papers/07cf7b3c7dc9037ad19f0b2ae81cde14

Thank you!
