### 2020-02-07 #1843 @jedbrown
I think it'll be straightforward or automatic upon acceptance -- I can't change that item on the website, but our EiC can when we accept (if it isn't automatic).

@ProbShin How is your review coming? Is there anything we can do to facilitate?
