### 2020-03-11 #1843 @ProbShin
@shah314 Sorry for the late, I had been trapped by some accidents before. I am working on the reviews recently. Here are some [comments](https://github.com/shah314/graphcoloring/issues/5#issue-578970792) about the experiment results.
