### 2020-03-11 shah314/graphcoloring#5 @shah314
@ProbShin @jedbrown

- I have added the number of edges in the git README.md. There is no space in the paper, so I left this information out of the paper.
- The found best and found worst are from running the algorithm 10 times with a different random seed.
- The ColPack comparison was done on a few randomly chosen graphs. I think it is sufficient to compare the algorithms on these 9 graphs. If you think more comparison is required, I can add more graphs.
- I have added a geometric mean as the last row in the table in the paper.

Thanks for the detailed review!
