### 2020-03-11 shah314/graphcoloring#5 @jedbrown
We ask Whedon (the JOSS bot) to rebuild the PDF in the review thread: https://github.com/openjournals/joss-reviews/issues/1843#issuecomment-597437061
