### 2020-03-11 #1843 @ProbShin
@shah314  Thanks for your understanding.

Here are some [comments](https://github.com/shah314/graphcoloring/issues/6#issue-579422802) about the code implementation. And I need to emphasize again that some modification is recommended, and some modification is only needed depending on the goal of the project.
