### 2020-03-12 a507390
Add verbose option
Add verbose option
