+++ APICallHeuristics.java
@@ -1,3 +1,4 @@
+
 import com.gcol.GraphColoring;
 import com.gcol.GraphReader;
 import com.gcol.Graph;
@@ -6,13 +7,13 @@ public class APICallHeuristics
 {
 	public static void main(String [] args) throws Exception
 	{
-		Graph graph = GraphReader.readGraph("data.col");
+		Graph graph = GraphReader.readGraph("filename");
 		// where filename is a DIMACS formatted graph file
 		boolean localsearch = true;
 		int iteratedgreedyiterations = 1000;
 		int localsearchiterations = 1000;
 		int localsearchtimeinmilliseconds = 1000;
-		int [] colors = GraphColoring.colorGraph(graph, localsearch, iteratedgreedyiterations, localsearchiterations, localsearchtimeinmilliseconds);
+		int [] colors = GraphColoring.colorGraph(graph, localsearch, iteratedgreedyiterations, localsearchiterations, localsearchtimeinmilliseconds, false);
 
 		int maxColor = -1;
 		for(int i=0; i<colors.length; i++)
@@ -31,7 +32,7 @@ public class APICallHeuristics
 		System.out.println("Colors of Vertices: ");
 		for(int i=0; i<colors.length; i++)
 		{
-		    System.out.print((graph.nodes[i].value+1) + ":" + colors[i] + ",");
+		    System.out.print(colors[i] + " ");
 		}
 
 	}
