### 2020-03-12 shah314/graphcoloring#6 @shah314
@ProbShin @jedbrown

- I have added a verbose option to both Backtracking.java and GraphColoring.java (heuristic). But I think some basic information, like what coloring was found should be still outputted.

- The KNOWN_K is actually a lower bound. If known, the backtracking algorithm will spend time on coloring the graph with this chromatic number. It is initialized to the clique size which is a lower bound on the chromatic number of the graph. The algorithm iteratively increments this number and tries to color the graph with this incremented number of colors.

- I have removed the adjacency matrix from the graph. This was only used by the clique algorithm and was not absolutely essential to computing the clique.

Let me know your thoughts. Thank you!
