### 2020-03-12 shah314/graphcoloring#6 @ProbShin
Great.

- Yes, you are right. the number of colors and the color of the vertex are both the output of the algorithm. Since the vertex colors have been stored in the `class Graph :: class Node`, the number of colors could be output in an integer variable.  *Print* the relative information to the screen can be considered as a debug operation by the users or the developers since this function is independent of the algorithms.

- I see. `KNOWN_K` is the lower bound. Thanks leading me out. And the clique size is a lower bound of the chromatic number. But unfortunately, there is not an efficient way to get this number out in constant or linear time for general graphs.

- Good, that's also my observation, it seems that the *adjacency matrix* is not used within the function of colorings.

In all, I am fine with all the modifications. Thanks.
