### 2020-03-13 #1843 @jedbrown
@ProbShin Thanks for your review. Before we proceed, could you please you check off your checklist at [the top of this thread](https://github.com/openjournals/joss-reviews/issues/1843#issue-514116573)?
