### 2020-03-13 #1843 @ProbShin
@jedbrown Sure, I tried, but I cannot edit the checklist. I noticed that the invitation link might be revoked. Is it possible to reinvite me somehow, I guess that might grant me access to the checklist.
