+++ src/main/java/com/gcol/GraphColoring.java
@@ -46,7 +46,7 @@ public class GraphColoring {
     	Option o7 = new Option("h", false, "This help message");
     	o7.setRequired(false);
     	options.addOption(o7);
-    	Option o8 = new Option("v", true, "Verbose");
+    	Option o8 = new Option("v", true, "Verbose true/false");
     	o8.setRequired(false);
     	options.addOption(o8);
     	
