### 2020-04-02 #1843 @arfon
@shah314 - could you update the author list on your Zenodo record to match the paper please? It currently includes @jedbrown which I'm guessing is due to him making a few edits to the paper.
