### 2020-04-03 #1843 @arfon
@shah314 - I made some minor changes to your paper in https://github.com/shah314/graphcoloring/pull/7 that improve the readability.
